+++ packages/core/src/magicswap/pools.ts
